got it to use headers.
Having trouble getting the content type header to set to application/json.  It either causes the server to stop registering it as a valid request, or it stops chrome sending. Possibly it's just not a feature of native JS, and I'd need to use AJAX. In any case: now it sends with headers